IGNAS_requirements.txt
absl-py
astunparse
biopython
cachetools
certifi
chardet
cycler
Cython
dataclasses
flatbuffers
future
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
importlib-metadata
joblib
Keras-Preprocessing
kiwisolver
Markdown
matplotlib
numpy
oauthlib
opt-einsum
pandas
Pillow
protobuf
pyasn1
pyasn1-modules
pyparsing
pyteomics
python-dateutil
pytz
requests
requests-oauthlib
rsa
scikit-learn
scipy
six
sklearn
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
threadpoolctl
torch
typing-extensions
UpSetPlot
urllib3
Werkzeug
wrapt
zipp
